test: Relocate size regression guards to module-level, add documentation
Addresses PR #196 review feedback by organizing tests and documentation to align
with the boxing strategy changes from #194-#195.

Changes:
- Move size regression guard tests from integration test file to their respective
  type modules (EngineOutput, ActionOutput, GenerateAlgoOrdersOutput) as
  module-level #[cfg(test)] mod size_guard blocks for tighter coupling
- Add detailed documentation explaining item-level boxing strategy in
  SendRequestsOutput (keeps empty case allocation-free vs field-level boxing)
- Add caveat to audit_seam benchmark documenting single-thread runtime overhead
  (A/B may overstate send cost vs multi-threaded production deployment)
- Add Item-only invariant assertion in market_data_from_file helper to catch
  future fixtures with Reconnecting events that would be silently reordered
- Fix backtest bench setup panic by sorting market_data_from_file input
  ascending by time_exchange before construction (fixture was interleaved, not
  globally sorted)
- Add CHANGELOG entry documenting backtest setup fix with compatibility note
